requirements.txt
fastapi
uvicorn
langchain
langchain_google_genai
langchain_openai
langchain_core
langchain_community
httpx-oauth
PyJWT
chromadb
firebase_admin
tiktoken
pymongo